requirements.txt
Flask
werkzeug==2.2.2
Flask-Cors
Flask-Session
oauthlib
python-dotenv
markdown



